+++ database/migrations/2021_03_09_191554_create_product_characteristics_table.php
@@ -16,7 +16,7 @@ class CreateProductCharacteristicsTable extends Migration
         Schema::create('product_characteristics', function (Blueprint $table) {
             $table->id();
             $table->bigInteger('product_id')->unsigned()->index();
-            $table->bigInteger('category_characteristic_id')->unsigned()->index();
+            $table->integer('category_characteristic_id')->unsigned()->index();
             $table->tinyInteger('val_boolean')->default(0)->index();
             $table->float('val_numeric')->default(0)->index();
             $table->string('val_short_text',512)->default('')->index();
